+++ src/util.ts
@@ -1,0 +1,28 @@
+
+'use strict';
+
+import fs = require('fs');
+
+export function semVer(a: string, b: string): number {
+    a = a.split(' ')[1];
+    b = b.split(' ')[1];
+    var pa = a.split('.');
+    var pb = b.split('.');
+    for (var i = 0; i < 3; i++) {
+        var na = Number(pa[i]);
+        var nb = Number(pb[i]);
+        if (na > nb) { return 1; }
+        if (nb > na) { return -1; }
+        if (!isNaN(na) && isNaN(nb)) { return 1; }
+        if (isNaN(na) && !isNaN(nb)) { return -1; }
+    }
+    return 0;
+}
+
+export function fileExists(filePath: string): boolean {
+	try {
+		return fs.statSync(filePath).isFile();
+	} catch (e) {
+		return false;
+	}
+}